requirements.txt
streamlit
google-genai
pandas
psycopg[binary]
python-dotenv